feature-#77 [계산기] 버그 수정 및 계량기 완성

+++ app/src/main/java/cookcook/nexters/com/amoogye/views/tools/add_tools/AddUtilActivity.kt
@@ -221,7 +221,7 @@ class AddUtilActivity : AppCompatActivity(), OnEditTextClickListener,
         newItem.unitNameBold = MeasureUnitSaveData.getInstance().unitNameBold
         newItem.unitNameSoft = MeasureUnitSaveData.getInstance().unitNameSoft
         newItem.unit = MeasureUnitSaveData.getInstance().unit
-        newItem.unitValue = MeasureUnitSaveData.getInstance().unitValue
+        newItem.unitValue = MeasureUnitSaveData.getInstance().unitValue.toDouble()
 
         realm.commitTransaction()
     }
